collapse menu at start and on resize
What happened here is that whole header is no longer under
the controller of the page, and thus somebody somewhere else
but the Controller needs to set the isCollapsed to true at
start. Since all menu directives share scope (except mnMenuGroup),
it could have been any of them, but mnNavbarCollapsible makes
a logical candidate.

In addition, added resize handler to do the same, so on any
resizing menu is collapsed, thus not causing problems if
crossing the hamburger boundary width.

Note: I would prefer .directive to always be in a new line, so
that a change on one directive, if many are chained, is not
showing the other directives at all, when using git diff or
similar tools.

Note: I would really prefer using single command injection
array rather than separate $inject. It is unneccessary height
in the code file.

Signed-off-by: zeljko <zeljko@z-ware.fi>