dry up step two - functionize
A funciton that returns a function is a beauty of JavaScript
and should be used indiscriminately :) Here however it really
removes a bunch of repetitive code.

Intentionally the function is in the same file, so thatit can be
decided later where it should move.

Strangelly, the route is now wrong, instead of /#/activities it
is /activities, jumping our of angular app. When /#/ was added
it becomes URL encoded, which is also weird, but it works. this
needs looking into, but we want /activities to be back end JSON
restful interface, so this should be behind a hash.

Signed-off-by: zeljko <zeljko@z-ware.fi>