dry up step 3 - no repeat code at all
Once the functions are moved to mohican and sprocket directives
are removed, only one line will exist in this file doing
needed. Still if we want to override something fro mthe normal
behaviour it can be done with a second parameter to the define
function which can take an object with all parameters that can
override defaults if provided.

Signed-off-by: zeljko <zeljko@z-ware.fi>