eager loading, fucking fantastic
This is simply the best commit I have ever made!

Here is what is going on. The service loads firstFetchSize or records,
which should be set to a number larger than the page size, but of size
that is fast processed by both the back end and front end loading. In
existing setup, 200 records are loaded and 20 are returned to show in
the page.

Even before the page records are returned, another fetchSize count
of records is requested, records that succees those already got, and when
those arrive, another set of preceeding record is requested, until all
records of the current filter are received.

This way records are received in background while they are shown as soon
as some 'significant minimum' is received.

There is several steps still pending:

* Seriouv testing of marginal cases. It might be some offsetting is
  messed up. Also, need to test what happens when you interrupt loading
  by changing filter, for instance.
* First load should not really be the page and beyond, but around the
  page, which would help if user immediatelly moves to the previous page.
  It should not be assumed that page 1 is loaded first, on the contrary.
* Of course, service should be generalized, and become mohican, and any
  reference to 'activity' should be in extended object only.
* Finally, we should think what happens if you ask for a page that has not
  yet been loaded. It might be so that wizest (sic) is just to wait, as
  loading is still fast.
* Also important is to limit how many documents would in end be laoded.
  3070 in this test case is not a lot for Chrome on a storng machine, but
  there could be many more. Then we should stop before full filter is loaded.

But nevertheless, this is fucking fantastic commit!

Signed-off-by: zeljko <zeljko@z-ware.fi>