next: scope is what you want
It takes a bit of time to understand, but everything you
want to use from a controller should be in scope, not actually
in the controller. But hat said, see the next commit.

So you were somewhat right that Angular doesn't like you touching
global namespace, but it does let you change your own scope, which
IS your namespace while you are executing.

However, that $scope does not come form window, so you cannot use
window's variables, that is true. Just remember, add stuff to scope,
not to controller.

Signed-off-by: zeljko <zeljko@z-ware.fi>