step 3: do translation as 'deep' as possible
Here we ditch saying "hey translate this, hey translate that", but
we let mohican directive do the translation, so we can have simpler
and much drier templates. Remember: GOAL IS TO DRY UP THE APP, NOT
THE MOHICAN! Because Mohican we make once, templates we make
until we retire (so next 3-4 years if we do this right).

To achieve this, first we need to inject the mnTranslation to the
directive, second, we need to do the translation, and pass translated
mnText to the template.

Finally, notice that .directive('mnMenuGroupItem... is in the new line,
that is so that git diff does not show it as changed, only directive
before that one was changed. We should always try to put new line before
dot in cases like this (chained definitions).

Converselly, if controller is not really big, it can be defined inside
the link directive function, just like the link function.

Signed-off-by: zeljko <zeljko@z-ware.fi>