step 4: complete step 3 also for transcluded text
This is the same thing, but as transcluded test is treated as
an array of elements (but we know it is just text), we need a
little bit different (and more fucked up) approach.

But this gives us a chance to learn how the transclusion can be
modified. Because of that fifth parameter to the link function I
have decided to use Angular for UI wherever, whenever. Using that
I can implement my own little local webkit!

Signed-off-by: zeljko <zeljko@z-ware.fi>