finally: no more special directive, just template
Now we notice that our directive is called in only one place and
does really nothing but serve the template, and for that we have
ngInclude. Note that ngInclude, just like our late directive creates
a scope, so even that would not change.

Note the "'url'" syntax, that is a tricky thing required as Angular
will eval the content of the parameter, and this is very common behaviour.

Signed-off-by: zeljko <zeljko@z-ware.fi>