even drier: get read of repeated words and the 'base.'
The sref parameter is already optional, so we can use that
to populate it ourselves form the transcluded text if not give, for
case when (not-localized) label is the same as the route.

This qould be supersimple except that ui-sref requires a static
value and our link function triggers too late. However, we can use
something called pre-linking function. The link element of a directive
can either be a function, which is remembered by Angular as
the postlink, or simply link function, but it can also be an object with
two elements called pre and post, latter being the link function, but
the first being the pre-link function, which executes before ANY
link function.

There are limitations with what pre-link can do, but in our case it
works just fine. I hope.

Signed-off-by: zeljko <zeljko@z-ware.fi>