fill hate (sic): solving the jumpy footer
This commit only temporarily fixes the footer that jumps. The problem
was in the offset that it is given, which for some reason is based
on some element whose top actually moves for some reason.

This is only part of the problem, as there is at least fillHeight and
fullHeight, and only one should exist, plus it seems neither is done well.

In addition, while the first documents are being loaded, pager
shows only << and >>, which is a bit silly, so that is changed too.

Signed-off-by: zeljko <zeljko@z-ware.fi>