fix: remove page=1 if on that page
This commit also shows how other default parameters can be removeed
with a (almost) oneliner.

What hurts me, and what I want fixed, is the fact that injection is
done so far from definition! I really want injection array to be next
to the controller definition, which is in this case in
/routes/activities.js . Also, have in mind that some controllers might
want to inject some of their own things!

Injection array is just a normal array of strings, and that can be used for
our benefit, but it must be done simpler.

That all said, this is not a priority now, it is OK the way it is!

Signed-off-by: zeljko <zeljko@z-ware.fi>