stabilize the grid on exact width
The width is given by the backend, if it is too narrow, we just
chop off what doesn't fit. This meant that we had to also set
width on the inner documents.

The hope is that some day table can be replaced with something
else, such as simply divs, andmaybe some JS width management.

Signed-off-by: zeljko <zeljko@z-ware.fi>