optimizing: cache formatted value
This should ASAP move to the loading, same place where date is
parsed into date structure, so can the rest be. Create a field
"formatted" (to Ts) that contains all fields OF ALL LAYOUTS!
Because you do not know at this point when you would actually load
which layout.

The benefit of doing it there is that, although it slows down the
loading of the first page, the rest is presumably invisible for the
user. As well, there is no longer need for any function being called
here, but template can just render model.formatted[field.name], which
saves a stackframe creation.

Signed-off-by: zeljko <zeljko@z-ware.fi>

# Conflicts:
#	app/assets/javascripts/mohican/directives/mnViewField/controller.js