fix the last page showing
This is only a partial fix. There are more of them related
to this, and not all are in this service, but some are. Mainly
what needs to be decided is what happens if eager loading has
started, but now somebody requests a page that is rather far.
It does seem simpler at this point to just wait, if eager loading
has started, wait, you will get loaded!

But this fix is still part of that solution.

Signed-off-by: zeljko <zeljko@z-ware.fi>