fix double loading
This fix double loading of the controller! This was horribly
difficult to figure out even when I figured exact line to look at.

This is why I am not sure if === is really better to use than
the ==, but I must say that latter has issues with null/undefined.
Guess the only solution is TO BE CAREFUL!

Signed-off-by: zeljko <zeljko@z-ware.fi>