grid transclusion in item scope
This is a golden commit.

What is happening here is that whatever is trancluded in the mn-grid
element will be repeated in each row and have access to the element of
the row it is included in. This way we can make actions per row. Yet we
did our best to also keep access to the scope in which mn-grid itself
lives.

The example here with three icons shows the usage of item, which is the
looping variable of the row, and also the usage of ctrl, which is a
controller accessed via mn-grid's outer scope's ctrl member. As ctrl
member might not be defined in the outer scope, ctrlScope gives that
scope.

This whole inner linking actually does not take additional scope, but only
the outer, mn-grid and row's 'item scope', so it takes minimum time
required to render, as we would need to render it one way or the other.

There are some gotchas here, as unsuspecting programmer might be surprised
when she tries to access some variable that she can see int he current
controller but it is not already there - this is what Angular team has
tried to prevent, and we reverted that prevention, but it suites our
purpose. This can be prevented by always using controllerAs parameter!