refetch data from backend if backend filter changed
As the title said, that is what this commit does. It is a bit tricky
though as it might so be that the eager loading is still ongoing when
this happens. Bit of a releif is that we already have the 'one promise
at a time' policy, so there is only one HTTP request pending at a time,
which is both good for the back end and making this implementation much
easier.

Signed-off-by: zeljko <zeljko@zwr.fi>