removing magigal route_name + 'Service'
The problem with this is that a non-mn developer programmer has
no way of knowing how the service will be named. It is inside
the mn code that Service suffix is added. Thus I greatly dislike
the idea of that being used here, in the app part. This way user
can just say 'service' which is magic per ce but needed, and
there is some similarity from the service member definition
bellow.

Signed-off-by: zeljko <zeljko@z-ware.fi>