convention before configuration: dry grid
We set a convention that the 'owner' of the grid, so the controller
with parent (possibly grand-parent up to ten times, see the for loop)
must be the type that inherits from mnControllerBase (or provide
identical interface).

The owner optional parameter is exactly that (possibly grand-) parent
controller, but the parameter can still be provided in case we have more
than one grid on the page (which likely we might have in form views
where there are multiple detail tabs.

Once we have the owner, we can focus on the actual owner's fields where
earlier we have focused on parameters of the directive. We have not lost any
flexibility, however, if someone would want to use the diretive outside
Mohican base controller, she would need to implement such interface that
base controller provides.

With this convention inner mnGridActions becomes simpler to write, as
now we have the 'owner' of its direct parent, which is exactly
controller we were looking for.

This implementation is in fact at least not less efficient, as now there
is no two way binding (but direct reference) and thus no watching by
angular, we simply set the values directly up.

Signed-off-by: zeljko <zeljko@zwr.fi>