overlay for context menu
Added a full window overlay fully transparent that only closes
the context menu if clicked. For some reason header has some high
z-index, so this one has unbelievably high. Note that overlay is
not only transparent but also has cursor default so it does not
seem it is being clicke but effectively hides the context menu.

Signed-off-by: zeljko <zeljko@z-ware.fi>