context menu positioning, always visible
This is not a perfect solution, because context menu 'jumps' after
being shown, but is a very well working solution. What it does is
that it a bit too late (when context menu is already visible) checks
if it fits to the screen, and then it makes sure it does. It will not
work if context menu is bigger than grid half a grid in either
horizontal or vertical dimansion, but that is exteme case.

To make this nice and testable I have also created more creative
random menu.

Signed-off-by: zeljko <zeljko@z-ware.fi>