use mn- color scheme to define layout colors
The idea behind this commit is to define a minimum of colors that
are used accross the app and application developer can simply
pick different colors and replace them in only one place.

But also, the underlying logic is to define a constant such as
#123456 in only one place, never in two. Notice that still
individual usage of that color can be overriden elsewhere.

This commit contains also a small example fix: all colors should be
defined in config.css and never in mohican.css, so they can be
overriden!

And of course everything in config.css (as well as mohican.css)
should be defined with !default so it can be overriden.

Signed-off-by: zeljko <zeljko@z-ware.fi>