zeljko's too simple dialog
As our ng-view does not have a wrapping template, so we need to manually
add a sorcerer directive, which is basically invisible sitting there but
when called in the action it sprungs out some HTML wrapper with our given
template inside.

In an actual applications there could be more than one dialog - and this
should handle that case, but is not tested, it can easily be tested
via two options in the menu bar and two templates.

The use of templates is really good solution here as dialogs can, and
probably will be very long and putting them elsewhere might make ugly code.

Included CSS is very ugly but points out some positioning hacks to make
this doable. Still positioning needs to be properly done via JS.

Signed-off-by: zeljko <zeljko@z-ware.fi>