mn-dialog css: manage position based on height
This commit was way more difficult than previous.

What happens is again that we only know the width once we have rendered
everything, thus the timeout - as elements are rendered one by one,
height checking triggers way too often, triggering Angular's 'stack
overflow' limit.

As for calculation, we are using the same watch (it is not really possible
without a watch, but it is not very expensive one) that we used for menu.
This time we are trying to position the dialog into the middle of the
window, this we need to also be watching window size.

If dialog is not responsive (and needs to be carefully responsive to the
window object) it will just glue to left or top and cannot be scrolled.
So that basically means it has to be responsive.

Signed-off-by: zeljko <zeljko@z-ware.fi>