loading single document when form requests it
Controller should ask from a service a single document if it is
showing the form view, and service should promise to do it. If the
document requested had previously been eager laoded, service should
immediatlely return, but if it is not loaded, it should laod it as
a special case of 'backend filter'.

Signed-off-by: zeljko <zeljko@z-ware.fi>