form view for products, introducing mnid
Form view is straightforward, but pointed out at the need for
mn specific id, as Rails uses id and Mongoid _id and oid. Thus
we use mnid, but only on the client side, prepared before sending
by the server and used for finding ids.

It should be considered if mn should request this to be an id,
instead of leaving it to the application to decide.

Signed-off-by: zeljko <zeljko@z-ware.fi>