properly build orders and eager load from db
This small commit solves a lot of problems, and is both an ugly
hack as well as an engineering marvel. Mainly, it solves the
problem of loading and sending complete documents, including
fully embedded referenced documents.

The main problem stems from the fact that Order contains has_one,
has_many (both references) and embeds_many relations, where
each embedded document contains a has_many reloation. This causes
several N+1 problems in loading and assembling multiple documents.

Worth mentioning is that Order does not redundantly store any
related documents -and it should as changes in the referenced
documents should not cause changes to the document (user's name
change should not change the name under which she was noted in
the document).

Back to N+1 problem, Mongoid does support eager loading (in this
commit, eager laoding means eager loading from a database for
solving N+1 problem), but only one level deep, thus not being
able to eager load products referenced in the embedded order_items.

To solve this problem, related products' references are stored in
a helper collection, which is managed in the before_save callback,
and such helper is being cleaned before the document is being
sent to the client - new convention implies that all fields with
double underscore prefix are not sent (and single prefixed fields
are ignored when received from the client).

Finally, the build_object method gathers all the needed fields into
a single hash before jsonitizing it and sending to the client.

This commit show the power of eager loading, as even when the
DB is as fast as MongoDB on localhost, loading of 100 documents
can take a second or more.

References:

What is N+1:
  http://stackoverflow.com/q/97197/#97253
  http://guides.rubyonrails.org/active_record_querying.html#eager-loading-associations
12345678901234567890123456789012345678901234567890123456789012345678901234567890
Callbacks in Mongoid and ActiveRecord:
    http://mongoid.org/en/mongoid/docs/callbacks.html
    http://api.rubyonrails.org/classes/ActiveRecord/Callbacks.html
Mapping with index: http://stackoverflow.com/q/4697557#11280903

Signed-off-by: zeljko <zeljko@z-ware.fi>