order model: store redundant data with refs
Now redundant data is stored along referenced object for
both 1 to 1 and 1 to many relations. This basically removes
the need for pure collection of references as at least
some data is still needed to be stored with the master
document.

This also makes our system more query friend (as oposed to
update friendly) as there is no 'build object' function
when data is being sent to the client, but more work is
done when a document is created.

On update, redundant data is not automatcailly re-remembered.

Signed-off-by: zeljko <zeljko@z-ware.fi>