seeds made better: dates and options
This adds four fields, two for select (single compulsory optiona and
multiple where none selected is also allowed, and two for dates (one
is optional and can be null). Orders are seeded with some sort of logic.

This commit, as such, breaks the UI and UI needs to be fixed.

Signed-off-by: zeljko <zeljko@z-ware.fi>