editable production lines
This demo commit also contains multiple minor bug fixes. Remaining bug
is that the Mohican JS naming system does not like the underscore
linked words for names.

This commit first includes the Lines resource on the front end and
a bunch of changes on the back end to facilitate this. Among the biggest
are the calculated fields in the json sent to front and custom update.

Except for that, editing is now also possible from the dashboard,
each line has the edit icon.

Signed-off-by: zeljko <zeljko@z-ware.fi>