fix: on the fly authentication
This allows Devise Controller to receive JSON authentication
message. This actually sets all other controllers to receive both
HTML and JSON, but can be overriden in those controllers that do
not need it. But for now it works fine.

To kill own session execute following code in Rails Console (these
instructions correspon to current seeds):

    x = User.find_by email: 'mohican@zwr.fi'
    x.password = 'mohican1234'
    x.save
    # Session is dead, need to authenticate
    x.password = 'mohican123'
    x.save
    # Session is dead again, need to authenticate

Signed-off-by: zeljko <zeljko@z-ware.fi>