clean bundling fixed, rubocop updated with annoying parser
It turned out that clean bundling would not be able to find source
specified in the Mohican gem, so explicitelly specifying that source
and the gem from it in the app is a simple fix, but there must be
another one.

This commit also updates the Rubocop, which brings in parser gem
that will now every time print annoyig unneccessary warning message.

Signed-off-by: zeljko <zeljko@z-ware.fi>