fix double click by using $location
First a question: should this be vm.owner.resourceName or
some other variable, to get correct route base?

This commit relies on the $location, which is way more reliable
than the ui-router's transitionTo. In general, whenever the
controller is to be reloaded, transitionTo should be avoided
and this approach used instead. The drawback is that is is
'uglier' and it is close binding URL string to state, but as long
as that is agreed (and why not?) this method solves all problems.

The original problem is caused by ui-router's inflexibilness to any
use but the strict one, so basically notify: false has some side
effects. So recommendation at this point is to always use transitionTo
with notify:false, and if notify:true should be used, use the method
from this commit.

Note: this always clears the search path (there is no ?x=y in the
URL), but this same emthod can be easily used to add/remove/modify
search path as $location.search() returns a hash and accepts a hash,
so adding one more search criteria can also easily fit into two lines.

Signed-off-by: zeljko <zeljko@z-ware.fi>